feat: Create GameBoard component and move all game content from App into component
